+++ src/utils/index.ts
@@ -2,3 +2,5 @@ export * from './register-manager';
 export * from './update-manager';
 export * from './group-helpers';
 export * from './get-layout-patches';
+export * from './events';
+export * from './enrichLayoutWithDefaults';
